test: move cypress constants to one file
