+++ cypress/integration/native-uploady/NativeUploady-simple-spec.js
@@ -1,5 +1,5 @@
 import uploadFile from "../uploadFile";
-import { WAIT_SHORT } from "../specWaitTimes";
+import { WAIT_SHORT } from "../../constants";
 
 describe("NativeUploady - Simple", () => {
     const fileName = "flower.jpg";
