+++ cypress/integration/retry-hooks/RetryHooks-withRetry-batch-spec.js
@@ -1,6 +1,6 @@
 import uploadFile from "../uploadFile";
-import { BATCH_ADD, ITEM_ABORT, ITEM_FINISH } from "../storyLogPatterns";
-import { WAIT_SHORT } from "../specWaitTimes";
+import { BATCH_ADD, ITEM_ABORT, ITEM_FINISH } from "../../constants";
+import { WAIT_SHORT } from
 
 describe("RetryHooks - Retry Upload", () => {
     const fileName = "flower.jpg",
