--- cypress/integration/specWaitTimes.js
@@ -1,9 +1,0 @@
-export const WAIT_X_SHORT = 200;
-
-export const WAIT_SHORT = 800;
-
-export const WAIT_MEDIUM = 1500;
-
-export const WAIT_LONG = 2500;
-
-export const WAIT_X_LONG = 3000;
