+++ cypress/integration/upload-button/UploadButton-custom-input-button-spec.js
@@ -1,6 +1,6 @@
 import intercept from "../intercept";
 import uploadFile from "../uploadFile";
-import { ITEM_FINISH } from "../storyLogPatterns";
+import { ITEM_FINISH } from "../../constants";
 
 describe("UploadButton - With Custom File Input And Button", () => {
     const fileName = "flower.jpg";
